pyproject.toml
[project]
name = "py-hamilton-step"
version = "0.1.0"
authors = [
    { name = "Bradley Bare", email = "bradleybare@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5,<3", "pyserial-asyncio>=0.6,<0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/py_hamilton_step"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.tasks]
fix_socket = "sudo chmod 777 /Users/birdbare/Parallels/venus_star.sock"
